requirements.txt
Flask==3.1.2
flask_cors==6.0.2
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.2.7
langchain_huggingface==1.2.0
numpy==2.4.2
Pillow==12.1.0
pydantic==2.12.5
pypdf==6.6.2
python-dotenv==1.2.1
unstructured==0.18.27
Werkzeug==3.1.5
